1. Quieted some compiler warnings in fp_Page.cpp 2. Fixed what has to be a bug in fl_DocLayout.cpp
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@4348 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/fmt/xp/fp_Page.cpp
@@ -199,8 +199,8 @@ void fp_Page::draw(dg_DrawArgs* pDA, bool bAlwaysUseWhiteBackground)
 		ScaleLayoutUnitsToScreen = (double)pDA->pG->getResolution() / UT_LAYOUT_UNITS;
  		UT_sint32 xmin = pDA->xoff;
   		UT_sint32 ymin = pDA->yoff;
-		UT_sint32 height = (UT_sint32) getHeightInLayoutUnits() * ScaleLayoutUnitsToScreen;
-		UT_sint32 width = (UT_sint32) getWidthInLayoutUnits() * ScaleLayoutUnitsToScreen ;
+		UT_sint32 height = (UT_sint32) ((double)getHeightInLayoutUnits() * ScaleLayoutUnitsToScreen);
+		UT_sint32 width = (UT_sint32) ((double)getWidthInLayoutUnits() * ScaleLayoutUnitsToScreen);
 		pDA->pG->fillRect(*pClr,xmin,ymin,width,height);
 	}
